split up ngrok code
