+++ packages/xdl/src/project/errors.ts
@@ -1,0 +1,7 @@
+import { XDLError } from '../xdl';
+
+export async function assertValidProjectRoot(projectRoot: string) {
+  if (!projectRoot) {
+    throw new XDLError('NO_PROJECT_ROOT', 'No project root specified');
+  }
+}
